+++ pkg/webscan/PrintEngine.go
@@ -1,0 +1,7 @@
+package webscan
+
+import "fmt"
+
+func (engine Engine) PrintEngine() {
+	fmt.Println(engine)
+}