+++ pkg/webscan/ScanDnsDetailed.go
@@ -20,7 +20,21 @@ func (engine Engine) ScanDnsDetailed(inputUrl string) (Engine, error) {
 		return engine, err
 	}
 
-	// TODO follow CNAME if exists
+	if len(engine.dnsScanResult.ARecords) == 0 && len(engine.dnsScanResult.AAAARecords) == 0 && engine.FollowRedirects { // If neither A nor AAAA records exist & redirects should be followed
+		engine.DnsScanEngine, err = engine.DnsScanEngine.GetCNAMERecord(inputUrl) // Retrieve CNAME record if exists
+		if err != nil {
+			return engine, err
+		}
+		if engine.DnsScanEngine.CNAMERecord != "" { // If CNAME record exists
+			if engine.Verbose {
+				fmt.Println("No A or AAAA records for", inputUrl, ". Following CNAME...")
+			}
+			engine, err = engine.ScanDnsSimple(engine.dnsScanResult.CNAMERecord) // Follow CNAME recursively, but only checking A, AAAA, and CNAMEs
+			if err != nil {
+				return engine, err
+			}
+		}
+	}
 
 	return engine, nil
 }